male opravy

+++ src/Enemy.java
@@ -72,7 +72,7 @@ public class Enemy {
             graphics.setColor(Color.RED);
             graphics.fillRect(x-15,y-25,30,5);
             graphics.setColor(Color.GREEN);
-            int barWidth=(int)(30*(health/(type==2 ? 3 : 2)));
+            int barWidth=30*(health)/(type==2 ? 3 : 2);
             graphics.fillRect(x-15,y-25,barWidth,5);
         }
     }
